--- src/redux/Cart.ts
@@ -1,14 +1,0 @@
-export interface CartItem {
-  readonly id: number;
-  price: number;
-}
-
-export interface CartState {
-  cartProducts: CartItem[];
-  totalPrice: number;
-}
-
-export const initialState: CartState = {
-  cartProducts: [],
-  totalPrice: 0,
-};
