+++ src/redux/reducers/rootReducer.ts
@@ -1,0 +1,8 @@
+import { combineReducers } from "@reduxjs/toolkit";
+import cartSlice from "./cartReducer";
+
+const rootReducer = combineReducers({
+  cart: cartSlice.reducer,
+});
+
+export default rootReducer;
