+++ src/redux/store.ts
@@ -1,9 +1,9 @@
 import { configureStore } from "@reduxjs/toolkit";
-import cartSlice from "./cartSlice";
+import rootReducer from "./reducers/rootReducer";
 
 export const store = configureStore({
   reducer: {
-    cart: cartSlice.reducer,
+    reducer: rootReducer,
   },
 });
